refactor: split codex app-server adapter concerns
